deleteaccount handler verifies user password

+++ backend/internal/controller/auth.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 	"pwdmanager_api/internal/auth"
 	"pwdmanager_api/internal/database"
+	"pwdmanager_api/internal/helpers"
 	"pwdmanager_api/pkg/models"
 
 	"github.com/gin-gonic/gin"
@@ -21,6 +22,11 @@ func DeleteAccount(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
+	if !helpers.CompareHashes(auth_input.Password, user.Password) {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid credentials"})
+		return
+	}
+
 	err = db.DeleteAccount(user)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
